Did tasks 01 and 02 (my_strdup, concat_params)

+++ lib/my/src/my_strlen.c
@@ -7,7 +7,7 @@
 
 #include "my.h"
 
-int my_strlen(char const *str)
+size_t my_strlen(char const *str)
 {
     int result = 0;
